Create 'dateToBytes()' as a module import.
