+++ modules/icao9303/DigitalSealV3.js
@@ -6,6 +6,7 @@ import { c40Encode } from "./utilities/c40-encode.js";
 import { c40Decode } from "./utilities/c40-decode.js";
 import { VDS_MAGIC } from "./utilities/vds-magic.js";
 import { VDS_SIGNATURE_MARKER } from "./utilities/vds-signature-marker.js";
+import { dateToBytes } from "./utilities/dateToBytes.js";
 
 /**
  * Stores properties and methods for ICAO 9303 visible digital seals (VDSs)
@@ -206,8 +207,8 @@ class DigitalSealV3 {
       this.identifierCode +
       this.certReference
     ));
-    output = output.concat(DigitalSeal.dateToBytes(this.issueDate));
-    output = output.concat(DigitalSeal.dateToBytes(this.signatureDate));
+    output = output.concat(dateToBytes(this.issueDate));
+    output = output.concat(dateToBytes(this.signatureDate));
     output.push(this.featureDefinition);
     output.push(this.typeCategory);
     return output;
